Finished edit-profile page.
All the fnctionalities are done, a extra has added, the profile photo can be dragged into the input.

fixed an error in the phone number inpus in all site, that prevented the capacity to introduce some countries phone numbers. Now all numbres are supported.

Added a magnifying glass icon for future search buttons.

Added a new functionality to dynamicCssClasses, min and max width and height, by putting a "i" for min and "a" for max before of the value number.

Added normalizeCamelCase.js, with two util funtions, one to tranform from camelCase  to natural lenguage words and one to capitalize the first letter in a string.