Updated the workspace page with an input to search for panels and a select to filter between the panels the user owns or joins.
A profile can now also be found by the user's nickname, in the url, now it is not necessary to remember the profile id to access the user's profile.

Updated all links to the user profile that use the id instead of the nickname, to make it easier for the user to remember the url.