Finished Columns panel type, Added pdf viewer and implemented panel settings.
All functionalities for columns type panels are done, with creation edit and delete options for columns and cards, two types of cards, text cards and pdf cards.
Panel settings added, only the panel creator user can modify it, theres options to add panel participants or delete them and an option to delete the current panel.

Implemented pdf viewer and pdf preview (that creates an image of the first page of the document to show it before enter to the document), the viewer allows the user to select the document text, and zoom in the document in three diferents ways, one with Ctrl + mouseWheel, other with two buttons to sum or substract and other throught an input, zoom cam be between 50% and 300% values.

Added a script that changes the color of the text and some relevant interface colors, to match with the background picture added by the user doing the text more easy to read.

Added in workspace a button to add a panel when the user already has at least one.