Added Explore page and fixed some errors
Implemented all the functionanilites of the Explore page, with the posibiliy to search panels and users, and a automatic random "feed", a button gives us the capacity to change from users to panels.

Added "subcriotionPanels" page, to show which plans are available before the users have an account.

fixed some errors with the "follow" button in the userProfile page, and an error that allows an user to move panel cards even if he was not invited to that panel.

Fixed the back button in panels, now it go back without errors.